Move Comparators and SortableRequest into api module
